refactor(entities): move entity platforms to entities/ package
Phase 2 of project structure reorganization (#625):
- Move sensor.py, binary_sensor.py, switch.py, number.py, select.py, button.py to entities/
- Update relative imports (.const -> ..const, .coordinator -> ..coordinator)
- Move test_sensor_state_class.py to tests/entities/
- All 700 tests pass
